tweak to start chrome browser window maximized as not all DOM shown if window to small + tweak README

+++ selenium-browser/src/main/java/com/xebialabs/xldeploy/stresstests/seleniumbrowser/Main.java
@@ -3,6 +3,7 @@ package com.xebialabs.xldeploy.stresstests.seleniumbrowser;
 import com.xebialabs.xldeploy.stresstests.seleniumbrowser.simulations.ImportApplication;
 import com.xebialabs.xldeploy.stresstests.seleniumbrowser.simulations.Login;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
 
 import java.io.File;
 import java.io.IOException;
@@ -15,28 +16,41 @@ public class Main {
 
     private static final String WEBDRIVER_CHROME_DRIVER_KEY = "webdriver.chrome.driver";
     private static final String WEBDRIVER_CHROME_DRIVER = "chromedriver";
+    private static final String TEMP_DIR = "java.io.tmpdir";
+    private static final String OS_MAC_DRIVER_KEY = "-mac";
+    private static final String OS_LINUX_DRIVER_KEY = "-linux";
+    private static final String OS_WIN_DRIVER_KEY = ".exe";
+
 
     public static void main(String[] args) {
-        ChromeDriver driver = new ChromeDriver();
-        System.setProperty(WEBDRIVER_CHROME_DRIVER_KEY, copyDriver(WEBDRIVER_CHROME_DRIVER));
+        String os = getOSDriverKey();
+        System.setProperty(WEBDRIVER_CHROME_DRIVER_KEY, copyDriver(WEBDRIVER_CHROME_DRIVER, os));
+        ChromeOptions options = new ChromeOptions();
+        if (OS_LINUX_DRIVER_KEY.equals(os) || OS_MAC_DRIVER_KEY.equals(os)) {
+            options.addArguments("--kiosk");
+        } else {
+            options.addArguments("--start-maximized");
+        }
+        ChromeDriver driver = new ChromeDriver(options);
 
         // login is the pre-requisite of each simulation scenario
         Login login = new Login();
         login.simulate(driver);
 
-        // other scenario's are given as command-line parameter
         ImportApplication packageUpload = new ImportApplication();
         packageUpload.simulate(driver);
     }
 
-    private static String copyDriver(String driverName) {
-        String OS = getOSName();
-        String tempDir = System.getProperty("java.io.tmpdir");
+    private static String copyDriver(String driverName, String os) {
+        String tempDir = System.getProperty(TEMP_DIR);
+        Path resolve = Paths.get(tempDir).resolve(driverName.concat(os));
+        if (resolve.toFile().exists()) {
+            return resolve.toString();
+        }
 
         try (
-            InputStream chromeDriver = Main.class.getClassLoader().getResourceAsStream(driverName.concat(OS)))
+            InputStream chromeDriver = Main.class.getClassLoader().getResourceAsStream(driverName.concat(os)))
         {
-            Path resolve = Paths.get(tempDir).resolve(driverName + OS);
             Files.copy(chromeDriver, resolve);
             File file = resolve.toFile();
             file.setExecutable(true);
@@ -47,20 +61,18 @@ public class Main {
         }
     }
 
-    private static String getOSName(){
-        String OS = System.getProperty("os.name").toLowerCase();
-        if (OS.indexOf("win") >= 0) {
-            OS = ".exe";
-        } else if (OS.indexOf("mac") >= 0) {
-            OS = "-mac";
-        } else if (OS.indexOf("nix") >= 0 || OS.indexOf("nux") >= 0 || OS.indexOf("aix") > 0
-                || OS.indexOf("sunos") >= 0) {
-            OS = "-linux";
+    private static String getOSDriverKey(){
+        String os = System.getProperty("os.name").toLowerCase();
+        if (os.indexOf("win") >= 0) {
+            return OS_WIN_DRIVER_KEY;
+        } else if (os.indexOf("mac") >= 0) {
+            return OS_MAC_DRIVER_KEY;
+        } else if (os.indexOf("nix") >= 0 || os.indexOf("nux") >= 0 || os.indexOf("aix") > 0
+                || os.indexOf("sunos") >= 0) {
+            return OS_LINUX_DRIVER_KEY;
         } else {
             throw new RuntimeException("OS type unsupported");
         }
-        return OS;
     }
-
 
 }
